chore: setup cloud environment
