+++ cloud/_tests_/unit/sample.test.js
@@ -1,0 +1,5 @@
+describe('Sample Test', () => {
+    it('should test that true === true', () => {
+      expect(true).toBe(true)
+    })
+  })
