+++ cloud/_utils_/database/migrate.js
@@ -4,7 +4,7 @@
 * 2019-05-01
 */
 
-var Parse = require("../utils/parse").getParse();
+var Parse = require("../parse").getParse();
 require("../migrations/classes/roles").migrate(Parse)
 .then(() => {
   require("../migrations/data/first-admin").setFirstAdmin(Parse);
