+++ cloud/src/classes/patient.model.js
@@ -1,0 +1,10 @@
+'use strict';
+
+class Patient {
+    constructor(Parse){
+        this.Parse = Parse;
+        this.ParseClass = "SurveyData";
+    }
+}
+
+export default Patient;
