+++ cloud/src/classes/vitals.model.js
@@ -1,0 +1,10 @@
+'use strict';
+
+class Vitals {
+    constructor(Parse){
+        this.Parse = Parse;
+        this.ParseClass = "Vitals";
+    }
+}
+
+export default Vitals;
