+++ package.json
@@ -7,23 +7,23 @@
     "test": "tests"
   },
   "dependencies": {
-    "concurrently": "^5.0.0",
-    "dotenv": "^5.0.1",
-    "events": "^2.0.0",
-    "express": "^4.16.3",
+    "concurrently": "5.0.0",
+    "dotenv": "5.0.1",
+    "events": "2.0.0",
+    "express": "4.16.3",
     "http": "0.0.0",
-    "json2csv": "^4.2.1",
-    "lodash": "^4.17.15",
-    "mailgun-js": "^0.22.0",
-    "moment": "^2.22.2",
-    "parse": "^1.11.1",
-    "parse-dashboard": "^2.0.0",
-    "parse-server": "^2.8.1",
-    "parse-server-simple-mailgun-adapter": "^1.0.1",
-    "path": "^0.12.7",
-    "request": "^2.88.0",
-    "request-promise": "^4.2.2",
-    "simple2e": "^1.0.2"
+    "json2csv": "4.2.1",
+    "lodash": "4.17.15",
+    "mailgun-js": "0.22.0",
+    "moment": "2.22.2",
+    "parse": "1.11.1",
+    "parse-dashboard": "2.0.0",
+    "parse-server": "2.8.1",
+    "parse-server-simple-mailgun-adapter": "1.0.1",
+    "path": "0.12.7",
+    "request": "2.88.0",
+    "request-promise": "4.2.2",
+    "simple2e": "1.0.2"
   },
   "devDependencies": {
     "case": "1.6.2",
@@ -34,23 +34,29 @@
     "fs": "0.0.1-security",
     "jasmine": "3.5.0",
     "jasmine-node": "1.16.2",
+    "jest": "25.2.4",
     "nodemon": "1.18.10",
     "pre-commit": "1.2.2",
     "randomstring": "1.1.5",
-    "standard-version": "7.0.0"
+    "standard-version": "7.0.0",
+    "supertest": "4.0.2"
   },
   "scripts": {
-    "test": "./node_modules/jasmine/bin/jasmine.js tests/**/*.spec.js",
-    "test:models": "npm run clearDB && npm run migrate && ./node_modules/jasmine/bin/jasmine.js tests/model-tests/**/*.spec.js",
-    "test:functions": "npm run clearDB && npm run migrate && ./node_modules/jasmine/bin/jasmine.js tests/function-tests/**/*.spec.js",
+    "test": "jest",
     "start": "node index.js",
-    "start-with-db": "concurrently \"mongod\" \"npm run start\"",
+    "start-with-db": "concurrently \"sudo mongod\" \"npm run start\"",
     "clearDB": "node scripts/clear-database.js",
     "migrate": "node scripts/migrate.js",
     "reinstall": "rm -rf node_modules/ && npm install",
     "gh-prune": "git fetch -p && for branch in `git branch -vv | grep ': gone]' | awk '{print $1}'`; do git branch -D $branch; done",
     "release": "standard-version",
     "snyk-protect": "snyk protect"
+  },
+  "jest": {
+    "testEnvironment": "node",
+    "coveragePathIgnorePatterns": [
+      "/node_modules/"
+    ]
   },
   "repository": {
     "type": "git",
