--- tests/index.spec.js
@@ -1,9 +1,0 @@
-var index = require("../index");
-var chai = require("chai");
-
-var expect = chai.expect;describe("index.js tests", function() {
-    it("addTwoNumbers returns a number", function() {
-        //expect(index.addTwoNumbers(0, 0)).to.be.a("number");
-        return 2
-    });
-});
